+++ infra/dependency/torch_dev.txt
@@ -1,2 +1,2 @@
 --extra-index-url "https://download.pytorch.org/whl/nightly/cpu"
-torch==2.7.0.dev20250209+cpu
+torch==2.8.0.dev20250408+cpu
