+++ infra/dependency/torchvision_dev.txt
@@ -1,2 +1,2 @@
 --extra-index-url "https://download.pytorch.org/whl/nightly/cpu"
-torchvision==0.22.0.dev20250209+cpu
+torchvision==0.22.0.dev20250408+cpu
